refactor, move common function from cmd/mesh to pkg
