+++ operator/pkg/manifest/shared_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package mesh
+package manifest
 
 import (
 	"bytes"
